RestAPI Project API testing using swagger
"This is a practice REST API project built using Spring Boot. It includes features like Spring Data JPA for seamless database interaction and MySQL for data storage. Swagger is integrated for comprehensive API documentation and testing, making it easier to explore and test endpoints. The project is configured with a production-ready profile and uses Postman for additional API testing. It serves as a foundation for learning and practicing RESTful service development."